+++ ChangeLog.md
@@ -4,6 +4,11 @@ T-Regx Changelog
 Incoming
 --------
 
+* Soon
+
+Added in 0.19.1
+---------------
+
 * Bug fixes
     * Fixed a bug when pattern `[]]` with prepared pattern was parsed incorrectly.
     * Fixed a bug when using character classes (e.g. `[:alpha:]`) with prepared pattern were parsed incorrectly.